@@ -21,7 +21,8 @@ import (
 	"path"
 	"time"
 
-	"golang.org/x/net/context"
+	"context"
+
 	appsv1api "k8s.io/api/apps/v1"
 	corev1api "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
